# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Clínica Médica — Sistema de Gestión de Turnos",
    "version": "1.0.0",
    "description": "Documentación completa del backend de la API REST para gestión de turnos médicos. Incluye módulos de Usuarios, Especialidades, Médicos, Obras Sociales, Pacientes, relación Médicos-Obras Sociales y Reservas de Turnos.\n\n**Autenticación:** El endpoint `/usuarios/login` devuelve un token JWT que debe enviarse en el header `Authorization: Bearer <token>` en los endpoints que lo requieran."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de desarrollo local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido desde el endpoint POST /usuarios/login"
      }
    },
    "schemas": {
      "SuccessListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {}
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "msg": {
            "type": "string",
            "example": "Descripción del error"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string"
                },
                "param": {
                  "type": "string"
                },
                "location": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "InternalServerError": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "ok": false,
              "msg": "Error interno del servidor"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Error de validación en los campos enviados",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/turnos/estadisticas": {
      "get": {
        "tags": ["Turnos"],
        "summary": "Obtener estadísticas de atenciones del mes anterior",
        "description": "Devuelve la cantidad de turnos y la facturación total agrupada por especialidad mediante un Stored Procedure.",
        "operationId": "getEstadisticas",
        "responses": {
          "200": {
            "description": "Estadísticas calculadas correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/turnos/reporte-pdf": {
      "get": {
        "tags": ["Turnos"],
        "summary": "Descargar informe general de turnos en PDF",
        "description": "Genera y descarga un archivo PDF con la auditoría de los turnos activos, fechas y obras sociales mediante pdfkit.",
        "operationId": "descargarReportePDF",
        "responses": {
          "200": {
            "description": "Archivo PDF generado y descargado con éxito"
          },
          "500": {
            "description": "Error al generar el informe en PDF"
          }
        }
      }
    },
    "/turnos/paciente/{id}": {
      "get": {
        "tags": ["Turnos"],
        "summary": "Obtener turnos de un paciente",
        "description": "Retorna los turnos activos de un paciente filtrando por `id_usuario`. Devuelve únicamente `fecha_hora` y `valor_total` de cada turno.",
        "operationId": "getPatientAppointments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del usuario (id_usuario) asociado al paciente",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turnos del paciente obtenidos con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fecha_hora": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-06-15T10:00:00.000Z"
                          },
                          "valor_total": {
                            "type": "number",
                            "format": "float",
                            "example": 3000.00
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "description": "No se encontraron turnos para el paciente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "ok": false,
                  "msg": "Turnos del paciente no encontrados"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/turnos/medico/{id}": {
      "get": {
        "tags": ["Turnos"],
        "summary": "Obtener turnos de un médico",
        "description": "Retorna los turnos activos de un médico filtrando por `id_usuario`. Devuelve únicamente `fecha_hora` y `valor_total` de cada turno.",
        "operationId": "getDoctorAppointments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del usuario (id_usuario) asociado al médico",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turnos del médico obtenidos con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fecha_hora": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-06-15T10:00:00.000Z"
                          },
                          "valor_total": {
                            "type": "number",
                            "format": "float",
                            "example": 5000.00
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "description": "No se encontraron turnos para el médico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "ok": false,
                  "msg": "Turnos del médico no encontrados"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}